Goodbye special strings!
When I started writing this kernel I wasn't a fan of C strings. It was always
problematic that you couldn't get their length easily. My idea was to embed the
length in the first byte of the string, and possibly extend it to the next
byte based on the top bit, and so on. I wrote some docs if you're curious.
Anyway, I'm not rebuilding the world, just a kernel, so this doesn't belong
here anymore.